+++ src/braindrop/app/screens/main.py
@@ -303,6 +303,8 @@ class Main(Screen[None]):
     def populate_display(self) -> None:
         """Populate the display."""
         self.query_one(Navigation).data = self._data
+        self.query_one(RaindropsView).data = self._data
+        self.query_one(RaindropDetails).data = self._data
         self.active_collection = self._data.all
         self.query_one(Navigation).highlight_collection(SpecialCollection.ALL())